Hotfix: Proper setup of env variables
The .env file is exported as environment variables in docker containers. Because of how DotEnv behaves, they are prioritized over any vars from other .env files, making it impossible to override them with .env.test for tests for example, except by making a dedicated container.

With this setup, APP_ENV and so on will never be part of the docker containers' environment variables, and will always be loaded by DotEnv from .env.* files, making it easier to override them.